requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
awkward==2.8.1
awkward_cpp==45
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cramjam==2.9.1
cycler==0.12.1
databricks-sdk==0.52.0
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
dgl==2.4.0+cu124
dill==0.4.0
docker==7.1.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fastjet==3.4.3.1
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
hepunits==2.3.6
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.35.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lightning-utilities==0.14.3
Mako==1.3.10
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mlflow==2.22.0
mlflow-skinny==2.22.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.3.2
multiprocess==0.70.18
MultiScaleDeformableAttention==1.0
narwhals==1.38.2
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging==24.2
pandas==2.2.3
parso==0.8.4
particle==0.25.3
pathos==0.3.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
pox==0.3.6
ppft==1.7.7
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.2
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
PyRoot==0.3.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.44.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.0
sympy==1.13.3
tbparse==0.0.9
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.4.0
torch-geometric==2.6.1
torchmetrics==1.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uproot==5.6.0
urllib3==2.3.0
uvicorn==0.34.2
vector==1.6.1
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
